refactor: extract execution helpers and SwapRequestHandler impl to execution.rs
Moves all swap execution logic (~1739 lines) out of lib.rs into a
dedicated execution.rs module, including pool interaction helpers,
DCA scheduling, swap batch processing, and the SwapRequestHandler
trait implementation.
